FRIDAY/requirements.txt
aiohttp==3.7.4.post0
async_timeout==3.0.1
beautifulsoup4==4.11.1
discord.py==1.7.3
Flask==2.0.1
Google_Images_Search==1.3.8
googletrans==3.0.0
gTTS==2.2.3
langcodes==3.1.0
praw==7.2.0
pytz==2021.1
random_topic==0.1.2
replit==3.2.4
requests==2.27.1
youtube_dl==2021.4.7
PyNaCl==1.5.0
language_data==1.1
Pillow==9.3.0
colour==0.1.5
random-topic==0.1.2
html5lib==1.1
chatbotAI==0.3.1.3
wikipedia==1.4.0
randomstuff.py==2.0.0
openai==0.10.1
python-dotenv==0.18.0
discord == 1.7.3
marisa-trie==0.7.8
language_data==0.0.1